URLManager添加

+++ HZExtend-Demo/HZExtend-Demo/Classes/Core/Model/HZModel.m
@@ -9,7 +9,7 @@
 #import "HZModel.h"
 #import "HZConst.h"
 #import "NSObject+HZExtend.h"
-
+#import "sqlite3.h"
 #import <objc/runtime.h>
 #import "FMDB.h"
 #define DBText  @"text"
@@ -495,7 +495,7 @@ static FMDatabase *DATA_BASE;
     
     NSMutableString *collection = [NSMutableString stringWithString:@"("];
     for (HZModel *model in array) {
-        [collection appendFormat:@"%ld,",model.primaryKey];
+        [collection appendFormat:@"%lu,",(unsigned long)model.primaryKey];
     }
     [collection deleteCharactersInRange:NSMakeRange(collection.length-1, 1)];
     [collection appendString:@")"];
